requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
build==0.10.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
ete3==3.1.3
executing==1.2.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip-tools==6.14.0
pipreqs==0.4.13
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
yarg==0.1.9
zipp==3.16.0
